+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [#141 changes to sapgui connections while connected](https://github.com/marcellourbani/vscode_abap_remote_fs/issues/141)
 - improved password deletion (cleanup cached value)
+- [#124 spaces not allowed in config name](https://github.com/marcellourbani/vscode_abap_remote_fs/issues/124)
 
 ## [0.10.2] 2021-03-28